requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
dataclasses-json==0.5.14
deprecation==2.1.0
Flask==2.3.3
frozenlist==1.4.0
google-api-core==2.11.1
google-auth==2.23.0
google-cloud-bigquery==3.11.4
google-cloud-core==2.3.3
google-cloud-documentai==2.18.0
google-cloud-documentai-toolbox==0.10.1a0
google-cloud-storage==2.10.0
google-cloud-vision==3.4.4
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.60.0
greenlet==2.0.2
grpc-google-iam-v1==0.12.6
grpcio==1.58.0
grpcio-status==1.58.0
idna==3.4
immutabledict==2.2.5
intervaltree==3.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
langchain==0.0.292
langsmith==0.0.37
lxml==4.9.3
MarkupSafe==2.1.3
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
natsort==8.4.0
numexpr==2.8.6
numpy==1.25.2
openai==0.28.0
packaging==23.1
pandas==2.1.0
pdftotext==2.
pikepdf==7.2.0
Pillow==9.5.0
proto-plus==1.22.3
protobuf==4.24.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.3.0
pydantic_core==2.6.3
PyMuPDF==1.23.3
PyMuPDFb==1.23.3
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
rsa==4.9
scikit-learn==1.2.2
scipy==1.11.2
six==1.16.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.20
tabulate==0.9.0
tenacity==8.2.3
threadpoolctl==3.2.0
tiktoken==0.5.1
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
Werkzeug==2.3.7
yarl==1.9.2
